harsat: Harmonized Regional Seas Assessment Tool

Tools for the assessment of data concerning contaminants (hazardous substances) and their effects in the marine environment. The code includes tools for pre-processing data, statistical trend analysis and comparison with threshold values, and post-processing for archiving and reporting.

See also

Author

Maintainer: OSPAR Commission (OSPAR) data@ospar.org [copyright holder, funder]

Authors:

Arctic Monitoring and Assessment Programme (AMAP) amap@amap.no [copyright holder, funder]

Helsinki Commission (HELCOM) secretariat@helcom.fi [copyright holder, funder]

International Council for the Exploration of the Sea (ICES) info@ices.dk [copyright holder]

AmbieSense Ltd info@ambiesense.com [copyright holder]

Other contributors:

HARSAT User Group harsat@ospar.org [contributor]