harsat: Harmonized Regional Seas Assessment Tool

Assessment of concentrations of hazardous substances and their biological effects in the marine environment. The code supports periodic international assessments by OSPAR, HELCOM and AMAP and also assessments by individual users. Includes tools for pre-processing data, statistical trend analysis and comparison with threshold values, and post-processing for archiving and reporting.

See also

Author

Maintainer: AmbieSense Ltd info@ambiesense.com [copyright holder]

Authors:

Rob Fryer rob.fryer@gov.scot

Leszek Kaliciak leszek@ambiesense.com

Scottish Government marinescotland@gov.scot [funder]

International Council for the Exploration of the Sea (ICES) info@ices.dk [copyright holder]

Other contributors:

Helsinki Commission (HELCOM) secretariat@helcom.fi [copyright holder, funder, contributor]

Arctic Monitoring and Assessment Programme (AMAP) amap@amap.no [copyright holder, funder, contributor]

OSPAR Commission (OSPAR) secretariat@ospar.org [copyright holder, funder, contributor]