Write assessment summary to a csv file

Creates a data frame summarising the assessment of each time series and writes it to a csv file. The summary includes:

meta-data such as the monitoring location and number of years of data for each time series

the fitted values in the last monitoring year with associated upper one-sided 95% confidence limits

the trend assessments (p-values and trend estimates)

the status assessments (if there any thresholds)

(optionally) a symbology summarising the trend (shape) and status (colour) of each time series

Usage

write_summary_table(
  assessment_obj,
  output_file = NULL,
  output_dir = ".",
  export = TRUE,
  determinandGroups = NULL,
  classColour = NULL,
  collapse_AC = NULL
)

Arguments

assessment_obj

An assessment object resulting from a call to run_assessment.

output_file

The name of the output csv file. If using NULL, the file will be called biota_summary.csv, sediment_summary.csv or water_summary.csv as appropriate. By default the file will be written to the working directory. If a file name is provided, a path to the output file can also be provided (e.g. using file.path). The output_dir option can also be used to specify the output file directory.

output_dir

The output directory for output_file. The default is the working directory. Any file path provided in output_file, will be appended to output_dir. The resulting output directory must already exist.

export

Logical. TRUE (the default) writes the summary table to a csv file. FALSE returns the summary table as an R object (and does not write to a csv file).

collapse_AC